Removed bugs in metadataparser caused by new format of csv file

+++ src/main/java/io/github/programminglife2016/pl1_2016/parser/metadata/SpecimenParser.java
@@ -89,7 +89,7 @@ public class SpecimenParser implements Parser {
             specimen.setAge(0);
         }
         else {
-            specimen.setAge((int) Double.parseDouble(string[1]));
+            specimen.setAge(Integer.parseInt(string[1]));
         }
         if (string[2].equals("Male")) {
             specimen.setMale(true);
